+++ paddle/fluid/primitive/codegen/gen.py
@@ -101,8 +101,10 @@ OTHER_PRIM_VJP_OPS = [
     'scatter_grad',
     'scatter_nd_add_grad',
     'slice_grad',
+    'squeeze_grad',
     'tile_grad',
     'topk_grad',
+    'unsqueeze_grad',
 ]
 
 # whole vjp list of primitive op vjp
